Day 9
By far the toughest challenge so far. I think I succeeded in using
modularity to sufficiently separate my methods (still room for
improvement) but unfortunately I accidentally/not so accidentally found
a hint for Part 2 on the subreddit. I originally thought I could get the
answer by running the same program but iterating through the characters
backwards, but I realized (with the hint) that I needed to recursively
expand the segments that were repeated until all of the compression
markers had been used. Interestingly, the answer was so long for Part 2
that the C# standard int (int32) was not large enough, causing my first
run of the program to give me a negative answer. Similarly interesting,
there was an extra character at the end (I think \n) that took me too
long to become aware of it's existence. A good challenge!