Finished!
Finished the last two puzzles and went back and completed the ones that
I hadn't been able to complete on my first attempt. Once I had wrapped
my head around how BFS applies to a number of the problems, I was able
to solve 13, 17, and 19 (and 11 eventually) fairly easily. Christmas is
saved!